requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fairscale==0.4.4
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.12.1
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycocoevalcap==1.2
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sacremoses==0.0.53
six==1.16.0
timm==0.4.12
tokenizers==0.10.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.15.0
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.14.0
